Move Locale Logic in Calculation to Separate Class
Make things more maintainable (I hope).
