+++ src/PhpSpreadsheet/Calculation/Token/Stack.php
@@ -4,6 +4,7 @@ namespace PhpOffice\PhpSpreadsheet\Calculation\Token;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
 use PhpOffice\PhpSpreadsheet\Calculation\Engine\BranchPruner;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
 
 class Stack
 {
@@ -12,7 +13,7 @@ class Stack
     /**
      * The parser stack for formulae.
      *
-     * @var mixed[]
+     * @var array<int, array>
      */
     private array $stack = [];
 
@@ -43,7 +44,7 @@ class Stack
         $this->stack[$this->count++] = $stackItem;
 
         if ($type === 'Function') {
-            $localeFunction = Calculation::localeFunc($value);
+            $localeFunction = Calculation::localeFunc(StringHelper::convertToString($value));
             if ($localeFunction != $value) {
                 $this->stack[($this->count - 1)]['localeValue'] = $localeFunction;
             }
